+++ src/main/java/com/wavers/server/QueryBase.java
@@ -1,0 +1,12 @@
+package com.wavers.server;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+
+public class QueryBase {
+    SQLiteDBManager dbManager = SQLiteDBManager.getInstance();
+
+    public Connection getConnection() throws SQLException {
+        return dbManager.getConnection();
+    }
+}
